LayoutTests/imported/w3c/IndexedDB-private-browsing/idbobjectstore_add12-expected.txt


PASS IDBObjectStore.add() - Attempt to add a record where the record's in-line key is not defined
